:sparkles: proto output mapper handler
