+++ gradle.properties
@@ -3,7 +3,7 @@ pluginGroup=com.guxingke.intellij
 pluginName=intellij-one
 
 # SemVer format -> https://semver.org
-pluginVersion=0.0.7
+pluginVersion=0.0.8
 
 # Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 pluginSinceBuild=222
